% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{aa_coi_PHMM}
\alias{aa_coi_PHMM}
\title{Amino acid profile hidden Markov model for coi5p.}
\format{
An object of class \code{PHMM} of length 14.
}
\usage{
aa_coi_PHMM
}
\description{
This model is stored in the coi5p package and was trained on a representative
sample of the barcode of life database (http://www.boldsystems.org/index.php).
}
\keyword{datasets}
