% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{aa_PHMM}
\alias{aa_PHMM}
\title{Amino acid profile hidden markov model for coi5p.}
\format{An object of class \code{PHMM} of length 14.}
\usage{
aa_PHMM
}
\description{
This model is stored in the coi5p package and was trained on a representitive
sample of the barcode of life database (http://www.boldsystems.org/index.php).
}
\keyword{internal}
