% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{example_barcode_data}
\alias{example_barcode_data}
\title{Example barcode data.
A nine line dataframe with the following columns:
id - the unique identifier for the sample
genetic_code - the genetic code for translation of the sample (features NA for unknowns)
taxa -  a taxonomic designation associated with the sample
sequence - the DNA sequence associated with the sample
notes - notes on the sequence structure}
\format{An object of class \code{data.frame} with 9 rows and 5 columns.}
\usage{
example_barcode_data
}
\description{
Example barcode data.
A nine line dataframe with the following columns:
id - the unique identifier for the sample
genetic_code - the genetic code for translation of the sample (features NA for unknowns)
taxa -  a taxonomic designation associated with the sample
sequence - the DNA sequence associated with the sample
notes - notes on the sequence structure
}
\keyword{datasets}
