% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{example_nt_string}
\alias{example_nt_string}
\title{Example coi5p DNA sequence string}
\format{
An object of class \code{character} of length 1.
}
\usage{
example_nt_string
}
\description{
This string of barcode data is used in the package documentation's examples
and within the vignette demonstrating how to use the package.
}
\keyword{datasets}
