% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi5p.r
\name{indel_check}
\alias{indel_check}
\alias{indel_check.coi5p}
\title{Check a translated coi5p sequence to see if an indel error is likely present}
\usage{
indel_check(x, ...)

\method{indel_check}{coi5p}(x, ..., indel_threshold = -346.95)
}
\arguments{
\item{x}{a coi5p class object for which frame() and translate() have been run.}

\item{indel_threshold}{}
}
\value{
an object of class code{"coi5p"}
}
\description{
Check a translated coi5p sequence to see if an indel error is likely present
}
\examples{

}
