% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_PHMMs.r
\name{individual_DNAbin}
\alias{individual_DNAbin}
\title{build an DNAbin with ape.
switches the - to n}
\usage{
individual_DNAbin(dna_string)
}
\description{
build an DNAbin with ape.
switches the - to n
}
\keyword{internal}
