% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_PHMMs.r
\name{ins_front_trim}
\alias{ins_front_trim}
\title{Check sequence for an early large string of deletions.
If it exists then return the starting index by which to slice the path and the string.}
\usage{
ins_front_trim(path_out, search_scope = 15)
}
\description{
Check sequence for an early large string of deletions.
If it exists then return the starting index by which to slice the path and the string.
}
\keyword{internal}
