% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.coi5p}
\alias{print.coi5p}
\title{Print a summary of a coi5p object.}
\usage{
\method{print}{coi5p}(x, ...)
}
\arguments{
\item{x}{a coi5p class object.}

\item{...}{additional arguments to be passed between methods.}
}
\description{
Print a summary of a coi5p object.
}
\keyword{internal}
