% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_PHMMs.r
\name{set_frame}
\alias{set_frame}
\title{Take an input sequence and get it into the reading frame.
Uses the path of the}
\usage{
set_frame(org_seq, path_out)
}
\description{
Take an input sequence and get it into the reading frame.
Uses the path of the
}
\keyword{internal}
