% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{trans_df}
\alias{trans_df}
\title{Data frame containing the translation table recommendation.}
\format{
An object of class \code{data.frame} with 4609 rows and 2 columns.
}
\usage{
trans_df
}
\description{
Data frame containing the translation table recommendation.
}
\keyword{datasets}
