% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi5p.r
\name{translate}
\alias{translate}
\alias{translate.coi5p}
\title{Translate a coi5p sequence.}
\usage{
translate(x, ...)

\method{translate}{coi5p}(x, ..., trans_table = 0, frame = 0)
}
\arguments{
\item{x}{a coi5p class object for which frame() has been run}

\item{frame}{}
}
\value{
an object of class code{"coi5p"}
}
\description{
Translate a coi5p sequence.
}
\examples{
#previously run functions:
dat = coi5p(example_nt_string )
dat = frame(dat)
#translate when the translation table is not known:
dat = translate(dat)
#translate when the translation table is known:
dat = translate(dat, trans_table = 5)
#additional components in output coi5p object:
data$aaSeq
}
\seealso{
\code{\link{coi5p}}

\code{\link{frame}}

\code{\link{which_trans_table}}
}
