% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.r
\name{translate_codon}
\alias{translate_codon}
\title{Censored Translation
Translate a DNA sequence using the censored translation table,
this translates codons for which the amino acids is unambigious across the
animal kingdom, and does not translate those for which the amino acid varies
but rather outputs a ? in the string}
\usage{
translate_codon(codon)
}
\description{
Censored Translation
Translate a DNA sequence using the censored translation table,
this translates codons for which the amino acids is unambigious across the
animal kingdom, and does not translate those for which the amino acid varies
but rather outputs a ? in the string
}
