% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.r
\name{which_trans_table}
\alias{which_trans_table}
\title{Determine the translation table to use for a given phylogenetic group.}
\usage{
which_trans_table(x)
}
\arguments{
\item{x}{a taxonomic designation (family order  class  phylum) first letter capitilizaed}
}
\value{
an integer indicating the correct translation table (give bcbi link here)
}
\description{
data stored down to family level.
relies on the above having been run so that the df is in the workspace and accessable
}
\examples{

}
