% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi5p.r
\name{new_coi5p}
\alias{new_coi5p}
\alias{validate_coi5p}
\alias{coi5p}
\title{Build a coi5p object from a DNA sequence string.}
\usage{
new_coi5p(x = character(), name = character())

validate_coi5p(new_instance)

coi5p(x = character(), name = character())
}
\arguments{
\item{x}{a nucleotide string.
Valid characters within the nucleotide string are: a,t,g,c,-,n.
The nucleotide string can be input as upper case, but will be automatically converted to lower case.}

\item{name}{an optional character string. Identifier for the sequence.}
}
\value{
an object of class code{"coi5p"}
}
\description{
Build a coi5p object from a DNA sequence string.
}
\examples{
dat = coi5p(example_nt_string)
#named coi5p sequence
dat = coi5p(example_nt_string, name = "example_seq1")
#components in output coi5p object:
data$raw
data$name
}
