% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi5p.r
\name{new_coi5p}
\alias{new_coi5p}
\alias{validate_coi5p}
\alias{coi5p}
\title{Build a coi5p object from a DNA sequence string}
\usage{
new_coi5p(x = character(), name = character())

validate_coi5p(new_instance)

coi5p(x = character(), name = character())
}
\arguments{
\item{x}{a nucleotide string.}

\item{name}{an optional character string. Identifier for the sequence.}
}
\value{
an object of class code{"coi5p"}
}
\description{
Build a coi5p object from a DNA sequence string
}
\examples{

}
