% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi5p.r
\name{frame}
\alias{frame}
\alias{frame.coi5p}
\title{Take a coi5p sequence and place it in reading frame.}
\usage{
frame(x, ...)

\method{frame}{coi5p}(x, ...)
}
\arguments{
\item{x}{a coi5p class object}

\item{...}{additional arguments to be passed between methods.}
}
\value{
an object of class \code{"coi5p"}
}
\description{
Take a coi5p sequence and place it in reading frame.
}
\examples{
#previously run function:
dat = coi5p(example_nt_string )

dat = frame(dat)

#additional components in output coi5p object:
dat$framed
}
\seealso{
\code{\link{coi5p}}
}
