% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{nt_PHMM}
\alias{nt_PHMM}
\title{Nucleotide profile hidden Markov model for coi5p.}
\format{An object of class \code{PHMM} of length 14.}
\usage{
nt_PHMM
}
\description{
This model is stored in the coi5p package and was trained on a representative
sample of the barcode of life database (http://www.boldsystems.org/index.php).
}
\keyword{internal}
