% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi5p.r
\name{translate}
\alias{translate}
\alias{translate.coi5p}
\title{Translate a coi5p sequence.}
\usage{
translate(x, ...)

\method{translate}{coi5p}(x, ..., trans_table = 0, frame = 0)
}
\arguments{
\item{x}{a coi5p class object for which frame() has been run}

\item{trans_table}{}
}
\value{
an object of class code{"coi5p"}
}
\description{
Translate a coi5p sequence.
}
\examples{

}
